vueのテストコードをtsへ移植 #188
ユーザーホームテスト
